refactor: move unrelated functions out of secrets.go
