+++ pkg/handlers/utils/owner_reference_test.go
@@ -1,4 +1,4 @@
-// Copyright 2024 Nutanix. All rights reserved.
+// Copyright 2025 Nutanix. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
 package utils
